Feat/author encounter creating

+++ src/Modules/Encounters/Explorer.Encounters.API/Dtos/EncounterDto.cs
@@ -3,14 +3,14 @@
     public class EncounterDto
     {
         public long AuthorId { get; set; }
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
         public int XP { get; set; }
         public string Status { get; set; }
         public string Type { get; set; }
-        public double Longitude { get; set; }
-        public double Latitude { get; set; }
+        public float Longitude { get; set; }
+        public float Latitude { get; set; }
         public double? LocationLongitude { get; set; }
         public double? LocationLatitude { get; set; }
         public string? Image { get; set; }
